+++ apps/main/src/app/api/crvusd/types.ts
@@ -1,6 +1,7 @@
+import type { NetworkEnum } from '@/loan/types/loan.types'
 import { Market } from '@curvefi/prices-api/crvusd'
 
 export type CrvUsdServerData = {
-  mintMarkets?: Market[]
+  mintMarkets?: Record<NetworkEnum, Market[]>
   dailyVolume?: number
 }
